config truncate (aon)

+++ application/controllers/bankname.php
@@ -59,8 +59,13 @@ class Bankname extends CI_Controller {
 	
 	function save(){
 		// ลบ receipt item ภายใต้ id ทั้งหมด
-		$this->db->where('1=1');
-		$this->db->delete('bnam');
+		if(db_helper_is_mssql($_this)){
+			$this->db->where('1=1');
+			$this->db->delete('bnam');
+		}
+		if(db_helper_is_mysql($_this)){
+			$this->db->truncate('bnam');
+		}
 
 		// เตรียมข้อมูล payment item
 		$bnam = $this->input->post('bnam');
