+++ application/controllers/material.php
@@ -417,8 +417,13 @@ class Material extends CI_Controller {
 		//$this->db->trans_start();  
 		
 		// ลบ receipt item ภายใต้ id ทั้งหมด
+		if(db_helper_is_mssql($_this)){
 		$this->db->where('1=1');
 		$this->db->delete('mtyp');
+		}
+		if(db_helper_is_mysql($_this)){
+		$this->db->truncate('mtyp');
+		}
 		//$this->db->delete('ktyp');
 
 		// เตรียมข้อมูล payment item
@@ -500,8 +505,13 @@ class Material extends CI_Controller {
 		//$this->db->trans_start();  
 		
 		// ลบ receipt item ภายใต้ id ทั้งหมด
+		if(db_helper_is_mssql($_this)){
 		$this->db->where('1=1');
 		$this->db->delete('mgrp');
+		}
+		if(db_helper_is_mysql($_this)){
+		$this->db->truncate('mgrp');
+		}
 		//$this->db->delete('ktyp');
 
 		// เตรียมข้อมูล payment item
